install.sh

+++ install.sh
@@ -20,21 +20,28 @@ echo "Mencari semua APK di GitHub Release..."
 
 APK_URLS=$(curl -s https://api.github.com/repos/$REPO/releases/tags/$TAG | grep browser_download_url | grep ".apk" | cut -d '"' -f 4)
 
+if [ -z "$APK_URLS" ]; then
+    echo "Tidak ada APK ditemukan di release."
+    exit 1
+fi
+
 echo ""
 echo "Downloading APK..."
 
 for url in $APK_URLS; do
     file=$(basename "$url")
     echo "Downloading $file"
-    wget -q --show-progress "$url"
+    wget --show-progress "$url"
 done
 
 echo ""
 echo "Installing APK..."
 
 for apk in *.apk; do
-    echo "Install $apk"
-    pm install -r -d "$apk"
+    if [ -f "$apk" ]; then
+        echo "Install $apk"
+        pm install -r -d "$apk"
+    fi
 done
 
 echo ""
